functions/chatcall/requirements.txt
slack_sdk
google-cloud-firestore
google-cloud-secret-manager
google_crc32c